Renamed access-om3 -> access-om3-cmake, access-om3-virtual -> access-om3
